+++ mbgdml/data/predictset.py
@@ -26,7 +26,7 @@
 import numpy as np
 from .. import __version__ as mbgdml_version
 from .basedata import mbGDMLData
-from ..predict import mbPredict
+from ..mbe import mbePredict, decomp_to_total
 
 
 class predictSet(mbGDMLData):
@@ -71,17 +71,24 @@ class predictSet(mbGDMLData):
         if pset is not None:
             self.load(pset)  
     
-    def prepare(
-        self, z, R, entity_ids, comp_ids, ignore_criteria=False
-    ):
+    def prepare(self):
         """Prepares a predict set by calculated the decomposed energy and
         force contributions.
-        """
 
-        self.E_decomp, self.F_decomp = self.predict.predict_decomposed(
-            z, R, entity_ids, comp_ids,
-            ignore_criteria=ignore_criteria, store_each=True
+        Note
+        -----
+        You must load a dataset first to specify ``z``, ``R``, ``entity_ids``,
+        and ``comp_ids``.
+        """
+        E_nbody, F_nbody, entity_combs, nbody_orders = self.mbePredict.predict_decomp(
+            self.z, self.R, self.entity_ids, self.comp_ids
         )
+        self.nbody_orders = nbody_orders
+        for i in range(len(nbody_orders)):
+            order = nbody_orders[i]
+            setattr(self, f'E_{order}', E_nbody[i])
+            setattr(self, f'F_{order}', F_nbody[i])
+            setattr(self, f'entity_combs_{order}', entity_combs[i])
         self._predicted = True
 
     def asdict(self):
@@ -102,13 +109,13 @@ class predictSet(mbGDMLData):
             'E_true': self.E_true,
             'e_unit': np.array(self.e_unit),
             'F_true': self.F_true,
-            'E_decomp': self.E_decomp,
-            'F_decomp': self.F_decomp,
             'entity_ids': self.entity_ids,
             'comp_ids': self.comp_ids,
-            'models_order': self.models_order,
+            'nbody_orders': self.nbody_orders,
             'models_md5': self.models_md5
         }
+        for nbody_order in self.nbody_orders:
+            pset[f'E_{nbody_order}'] = getattr(self, f'E_{nbody_order}')
 
         if self._loaded == False and self._predicted == False:
             if not hasattr(self, 'dataset') or not hasattr(self, 'mbgdml'):
@@ -163,7 +170,7 @@ class predictSet(mbGDMLData):
         Parameters
         ----------
         pset : :obj:`str` or :obj:`dict`
-            Path to predict set ``.npz`` file or a dictionary.
+            Path to predict set ``npz`` file or a dictionary.
         """
         if isinstance(pset, str):
             pset = dict(np.load(pset, allow_pickle=True))
@@ -175,21 +182,26 @@ class predictSet(mbGDMLData):
         self.version = str(pset['version'][()])
         self.z = pset['z']
         self.R = pset['R']
-        self.E_decomp = pset['E_decomp']
-        self.F_decomp = pset['F_decomp']
+
         self.r_unit = str(pset['r_unit'][()])
         self.e_unit = str(pset['e_unit'][()])
         self.E_true = pset['E_true']
         self.F_true = pset['F_true']
         self.entity_ids = pset['entity_ids']
         self.comp_ids = pset['comp_ids']
-        self.models_order = pset['models_order']
+        self.nbody_orders = pset['nbody_orders']
         self.models_md5 = pset['models_md5']
+
+        for i in pset['nbody_orders']:
+            setattr(self, f'E_{i}', pset[f'E_{i}'])
+            setattr(self, f'F_{i}', pset[f'F_{i}'])
+            setattr(self, f'entity_combs_{i}', pset[f'entity_combs_{i}'])
 
         self._loaded = True
 
-    def nbody_predictions(self, nbody_orders):
-        """Energies and forces of all structures .
+    def nbody_predictions(self, nbody_orders, n_workers=1):
+        """Energies and forces of all structures including ``nbody_order``
+        contributions.
 
         Predict sets have data that is broken down into many-body contributions.
         This function sums the many-body contributions up to the specified
@@ -198,8 +210,8 @@ class predictSet(mbGDMLData):
 
         Parameters
         ----------
-        nbody_orders : :obj:`list` [:obj:`int`]
-            N-body orders to include.
+        nbody_orders : :obj:`list` of :obj:`int`
+            :math:`n`-body orders to include.
         
         Returns
         -------
@@ -208,13 +220,22 @@ class predictSet(mbGDMLData):
         :obj:`numpy.ndarray`
             Forces of structures up to an including n-order corrections.
         """
+        if n_workers != 1:
+            global ray
+            import ray
+            ray.is_initialized()
+
         E = np.zeros(self.E_true.shape)
         F = np.zeros(self.F_true.shape)
         for nbody_order in nbody_orders:
-            E_nbody = [i[nbody_order]['T'] for i in self.E_decomp]
-            F_nbody = [i[nbody_order]['T'] for i in self.F_decomp]
-            E = np.add(E, E_nbody)
-            F = np.add(F, F_nbody)
+            E_decomp = getattr(self, f'E_{nbody_order}')
+            F_decomp = getattr(self, f'F_{nbody_order}')
+            entity_combs = getattr(self, f'entity_combs_{nbody_order}')
+            E_nbody, F_nbody = decomp_to_total(
+                E_decomp, F_decomp, self.entity_ids, entity_combs
+            )
+            E += E_nbody
+            F += F_nbody
         return E, F
 
     def load_dataset(self, dset):
@@ -260,7 +281,10 @@ class predictSet(mbGDMLData):
 
         self.z = dset['z']
         self.R = dset['R']
-        self.E_true = dset['E']
+        if isinstance(dset['E'], float):
+            self.E_true = np.array([dset['E']])
+        else:
+            self.E_true = dset['E']
         self.F_true = dset['F']
         self.entity_ids = dset['entity_ids']
         self.comp_ids = dset['comp_ids']
@@ -278,27 +302,50 @@ class predictSet(mbGDMLData):
         else:
             self.e_unit = dset['e_unit']
     
-    def load_models(self, models, use_torch=False):
+    def load_models(
+        self, models, predict_model, use_ray=False, n_cores=None,
+        wkr_chunk_size=100
+    ):
         """Loads model(s) in preparation to create a predict set.
 
         Parameters
         ----------
-        models : :obj:`list` of :obj:`str` or :obj:`dict`
-            GDML models in ascending order of n-body corrections (e.g., 1-, 2-
-            and 3-body models).
-        use_torch : :obj:`bool`, default: ``False``
-            Use PyTorch to make predictions.
+        models : :obj:`list` of :obj:`mbgdml.predict.mlWorker`
+            Machine learning model objects that contain all information to make
+            predictions using ``predict_model``.
+        predict_model : ``callable``
+            A function that takes ``z, r, entity_ids, nbody_gen, model`` and
+            computes energies and forces. This will be turned into a ray remote
+            function if ``use_ray = True``. This can return total properties
+            or all individual :math:`n`-body energies and forces.
+        use_ray : :obj:`bool`, default: ``False``
+            Parallelize predictions using ray. Note that initializing ray tasks
+            comes with some overhead and can make smaller computations much
+            slower. Thus, this is only recommended with more than 10 or so
+            entities.
+        n_cores : :obj:`int`, default: ``None``
+            Total number of cores available for predictions when using ray. If
+            ``None``, then this is determined by ``os.cpu_count()``.
+        wkr_chunk_size : :obj:`int`, default: ``100``
+            Number of :math:`n`-body structures to assign to each spawned
+            worker with ray.
         """
-        self.predict = mbPredict(models, use_torch=use_torch)
-
-        from mbgdml.data import mbModel
-        models_md5, models_order = [], []
-        for model in self.predict.models:
-            models_order.append(len(set(model['entity_ids'])))
+        if use_ray:
+            global ray
+            import ray
+            assert ray.is_initialized()
+        self.mbePredict = mbePredict(
+            models, predict_model, use_ray, n_cores, wkr_chunk_size
+        )
+        
+        models_md5, nbody_orders = [], []
+        for model in self.mbePredict.models:
+            if use_ray:
+                model = ray.get(model)
+            nbody_orders.append(model.nbody_order)
             try:
-                models_md5.append(model['md5'])
+                models_md5.append(model.md5)
             except AttributeError:
                 pass
-        self.models_order = np.array(models_order)
+        self.nbody_orders = np.array(nbody_orders)
         self.models_md5 = np.array(models_md5)
-